tlj  Solr完成

+++ store-shop-basic/store-stop-common/src/main/java/com/qf/service/ISolrService.java
@@ -3,5 +3,5 @@ package com.qf.service;
 import com.qf.dto.ResultBean;
 
 public interface ISolrService {
-    ResultBean seek(String keyword, Long page);
+    ResultBean seek(String keyword, Integer page);
 }
